move some exports to index.ts
