+++ src/client/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useOnClickOutside } from "./useOnClickOutside";
