+++ src/client/utils/findParentElement.ts
@@ -1,13 +1,13 @@
 export const findParentElement = <T extends HTMLElement>(
   tagName: string,
-  element: HTMLElement
+  element: HTMLElement | null
 ): T | null => {
+  if (!element) {
+    return null;
+  }
   if (element.tagName.toLowerCase() === tagName) {
     return element as T;
   }
-  if (element.parentElement) {
-    return findParentElement<T>(tagName, element.parentElement);
-  }
 
-  return null;
+  return findParentElement<T>(tagName, element.parentElement);
 };
