+++ tailwind.config.cjs
@@ -1,4 +1,11 @@
+// @ts-nocheck
 const colors = require("tailwindcss/colors");
+
+delete colors.lightBlue
+delete colors.warmGray
+delete colors.trueGray
+delete colors.coolGray
+delete colors.blueGray
 
 /** @type {import('tailwindcss').Config} */
 const config = {
